fix: 修复了删除用户上可能会出现的bug

+++ back/src/main/java/com/hongshu/service/impl/UserServiceImpl.java
@@ -165,6 +165,10 @@ public class UserServiceImpl implements UserService
     @Override
     public void deleteUser(Long id)
     {
+        if(id == UserUtils.getLoginUser().getId())
+        {
+            throw new MyException("不能删除正在登录的用户");
+        }
         userDao.deleteById(id);
     }
